Revert cluster monitoring
