+++ public/templates/management/management.pug
@@ -2,6 +2,7 @@ include ./management.head
 include ./welcome.html
 include ./status.html
 include ./configuration/configuration.pug
+include ./monitoring/monitoring.pug
 include ./logs.html
 include ./reporting.html
 include ./groups/groups.html
