--- public/templates/management/monitoring/monitoring.jade
@@ -1,7 +1,0 @@
-include ./monitoring.head
-include ./main.html
-include ./main-timelions.html
-include ./configuration.html
-include ./nodes-list.html
-include ./nodes-detail.html
-include ../../footer.foot
